+++ packages/storybook/src/react/Footer.stories.tsx
@@ -1,6 +1,7 @@
 import { Meta, StoryObj } from '@storybook/react-vite';
 import { Footer } from '@gemeente-denhaag/footer';
 import { contactData, footerLegalData, footerSocialData, newsletterData } from '../templates/util';
+import readme from '../../../../components/Footer/README.md?raw';
 
 const exampleArgs = {
   contactData: contactData,
@@ -13,6 +14,13 @@ const meta = {
   component: Footer,
   args: exampleArgs,
   tags: ['autodocs'],
+  parameters: {
+    docs: {
+      description: {
+        component: readme,
+      },
+    },
+  },
 } as Meta<typeof Footer>;
 
 export default meta;