+++ packages/storybook/src/react/UnorderedList.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Meta, StoryObj } from '@storybook/react-vite';
-import readme from '../../../../components/FormField/README.md?raw';
+import readme from '../../../../components/UnorderedList/README.md?raw';
 import { UnorderedList, UnorderedListItem } from '@gemeente-denhaag/unorderedlist';
 
 const meta = {